+++ common-utils-base/pom.xml
@@ -2,11 +2,11 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0"
          xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
     <parent>
         <groupId>com.el</groupId>
-        <artifactId>el-utils-parent</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../el-utils-parent</relativePath>
+        <artifactId>el-common-utils</artifactId>
+        <version>1.0.1-SNAPSHOT</version>
     </parent>
 
     <modelVersion>4.0.0</modelVersion>
