+++ common-utils-base/src/main/java/com/el/base/utils/support/exception/ExceptionHandlerManager.java
@@ -1,8 +1,6 @@
-package com.el.base.utils.support.exception.handler;
+package com.el.base.utils.support.exception;
 
-import com.el.base.utils.support.exception.AbstractDomainRuntimeException;
-import com.el.base.utils.support.exception.ExceptionHandler;
-import com.el.base.utils.support.exception.data.ErrorMessage;
+import com.el.base.utils.support.globalization.ErrorMessage;
 
 import java.util.Map;
 import java.util.Objects;
