+++ common-utils-core/pom.xml
@@ -4,11 +4,9 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <parent>
         <groupId>com.el</groupId>
-        <artifactId>el-utils-parent</artifactId>
-        <version>1.0.0-SNAPSHOT</version>
-        <relativePath>../el-utils-parent</relativePath>
+        <artifactId>el-common-utils</artifactId>
+        <version>1.0.1-SNAPSHOT</version>
     </parent>
-
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>common-utils-core</artifactId>
