+++ common-utils-core/src/main/java/com/el/common/web/access/filter/AbstractRouterFilterVerify.java
@@ -2,7 +2,6 @@ package com.el.common.web.access.filter;
 
 import com.alibaba.fastjson.JSON;
 import com.el.base.utils.support.exception.ExtendRuntimeException;
-import com.el.common.web.utils.TraceIdUtil;
 import com.el.common.web.config.CommonWebConstants;
 import com.el.common.web.access.FilterFailureType;
 import lombok.extern.slf4j.Slf4j;
